moving sidebar stories and changing paths
